refactor: name all files back to their original names
