+++ frontend/src/components/MainView/ColorToggle/ColorToggle.tsx
