+++ frontend/src/components/MainView/Legend/Legend.css
