+++ frontend/src/components/MainView/Legend/Legend.tsx
