+++ frontend/src/components/MainView/ZoomButtons/ZoomButtons.css
