+++ frontend/src/components/MainView/ZoomButtons/ZoomButtons.tsx
