+++ frontend/src/components/RoomModal/RoomModal.tsx
@@ -1,7 +1,7 @@
 import { X } from "lucide-react";
 import { useCallback, useRef, useState } from "react";
-import ConfirmationDialog from "../ConfirmationDialog";
-import RoomForm from "./RoomForm";
+import ConfirmationDialog from "../ConfirmationDialog/ConfirmationDialog";
+import RoomForm from "./RoomForm/RoomForm";
 import "./style.css";
 
 interface RoomModalProps {
