+++ frontend/src/components/Sidepanel/RoomInfo/RoomInfo.tsx
@@ -14,7 +14,7 @@ import { useRoomSelection } from "../../../hooks/useRoomSelection";
 import { editRoom } from "../../../services/roomsService";
 import { EXPAND_COLLAPSE_TRANSITION } from "../../../utils/motionTransitions";
 import { renderValue } from "../../../utils/renderValue";
-import RoomModal from "../../RoomModal";
+import RoomModal from "../../RoomModal/RoomModal";
 import "../style.css";
 
 type RoomInfoProps = {
