+++ frontend/src/components/Sidepanel/SidePanel.css
