+++ frontend/src/components/TopBar/TopBar.tsx
@@ -1,7 +1,7 @@
 import { Settings, User } from "lucide-react";
 import { useEffect, useState } from "react";
-import PersonSearch from "../PersonSearch";
-import SettingsModal from "./SettingsModal";
+import PersonSearch from "../PersonSearch/PersonSearch";
+import SettingsModal from "./SettingsModal/SettingsModal";
 import "./style.css";
 import TopBarMenu from "./TopBarMenu";
