+++ frontend/tests/components/personForm.test.tsx
@@ -2,7 +2,7 @@ import "@testing-library/jest-dom";
 import { fireEvent, render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { afterEach, describe, expect, it, vi } from "vitest";
-import PersonForm from "../../src/components/PersonModal/PersonForm";
+import PersonForm from "../../src/components/PersonModal/PersonForm/PersonForm";
 
 vi.mock("../../src/services/peopleService", () => ({
   findAllPeople: vi.fn().mockResolvedValue([
