+++ frontend/tests/components/personModal.test.tsx
@@ -7,7 +7,7 @@ import {
   within,
 } from "@testing-library/react";
 import { afterEach, describe, expect, it, vi } from "vitest";
-import PersonModal from "../../src/components/PersonModal";
+import PersonModal from "../../src/components/PersonModal/PersonModal.tsx";
 import { findAllDepartments } from "../../src/services/referenceDataService.ts";
 
 vi.mock("../../src/services/referenceDataService", () => ({
