+++ frontend/tests/components/roomForm.test.tsx
@@ -1,7 +1,7 @@
 import "@testing-library/jest-dom";
 import { fireEvent, render, screen, waitFor } from "@testing-library/react";
 import { afterEach, describe, expect, it, vi } from "vitest";
-import RoomForm from "../../src/components/RoomModal/RoomForm";
+import RoomForm from "../../src/components/RoomModal/RoomForm/RoomForm";
 
 vi.mock("../../src/services/referenceDataService", () => ({
   findAllDepartments: vi.fn().mockResolvedValue([
