+++ frontend/tests/components/topBar.test.tsx
@@ -2,7 +2,7 @@ import "@testing-library/jest-dom";
 import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { describe, expect, it, vi } from "vitest";
-import TopBar from "../../src/components/TopBar";
+import TopBar from "../../src/components/TopBar/TopBar";
 
 vi.mock("../../src/hooks/useAuth", () => ({
   useAuth: () => ({
